+++ src/App/NetDaemon.App/Common/Fluent/Fluent.cs
@@ -1,8 +1,5 @@
-﻿using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Concurrent;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace JoySoftware.HomeAssistant.NetDaemon.Common
@@ -35,6 +32,26 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
     }
 
     /// <summary>
+    ///     When state change
+    /// </summary>
+    public interface IDelayStateChange
+    {
+        /// <summary>
+        ///     When state change from or to a state
+        /// </summary>
+        /// <param name="to">The state change to, or null if any state</param>
+        /// <param name="from">The state changed from or null if any state</param>
+        /// <param name="allChanges">Get all changed, even only attribute changes</param>
+        IDelayResult DelayUntilStateChange(object? to = null, object? from = null, bool allChanges = false);
+
+        /// <summary>
+        ///     When state change, using a lambda expression
+        /// </summary>
+        /// <param name="stateFunc">The lambda expression used to track changes</param>
+        IDelayResult DelayUntilStateChange(Func<EntityState?, EntityState?, bool> stateFunc);
+    }
+
+    /// <summary>
     ///     Represents an entity
     /// </summary>
     public interface IEntity :
@@ -48,6 +65,11 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
     public interface IEntityProperties
     {
         /// <summary>
+        ///     Filter on area where the entity device are at
+        /// </summary>
+        string? Area { get; set; }
+
+        /// <summary>
         ///     Filter on attribute
         /// </summary>
         dynamic? Attribute { get; set; }
@@ -56,11 +78,6 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
         ///     Filter on unique id of the entity
         /// </summary>
         string EntityId { get; set; }
-
-        /// <summary>
-        ///     Filter on area where the entity device are at
-        /// </summary>
-        string? Area { get; set; }
 
         /// <summary>
         ///     Filter on last changed time
@@ -187,93 +204,12 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
     }
 
     /// <summary>
-    ///     When state change
-    /// </summary>
-    public interface IDelayStateChange
-    {
-        /// <summary>
-        ///     When state change from or to a state
-        /// </summary>
-        /// <param name="to">The state change to, or null if any state</param>
-        /// <param name="from">The state changed from or null if any state</param>
-        /// <param name="allChanges">Get all changed, even only attribute changes</param>
-        IDelayResult DelayUntilStateChange(object? to = null, object? from = null, bool allChanges = false);
-
-        /// <summary>
-        ///     When state change, using a lambda expression
-        /// </summary>
-        /// <param name="stateFunc">The lambda expression used to track changes</param>
-        IDelayResult DelayUntilStateChange(Func<EntityState?, EntityState?, bool> stateFunc);
-    }
-
-    /// <summary>
     ///     Track states on entities
     /// </summary>
     public interface IStateEntity : ITurnOff<IStateAction>, ITurnOn<IStateAction>, IToggle<IStateAction>,
         ISetState<IStateAction>
     {
     }
-
-    // /// <summary>
-    // ///     Time operations
-    // /// </summary>
-    // public interface ITime
-    // {
-    //     /// <summary>
-    //     ///     Run an action once every period of time
-    //     /// </summary>
-    //     /// <param name="timeSpan">The time between runs</param>
-    //     ITimeItems Every(TimeSpan timeSpan);
-    // }
-
-    // /// <summary>
-    // ///     Entities used on time actions
-    // /// </summary>
-    // public interface ITimeItems
-    // {
-    //     /// <summary>
-    //     ///     Use entities with lambda expression for further actions
-    //     /// </summary>
-    //     /// <param name="func">Lambda expression to filter out entities</param>
-    //     ITimerEntity Entities(Func<IEntityProperties, bool> func);
-
-    //     /// <summary>
-    //     ///     Use entities from list
-    //     /// </summary>
-    //     /// <param name="entities">The entities to perform actions on</param>
-    //     ITimerEntity Entities(IEnumerable<string> entities);
-
-    //     /// <summary>
-    //     ///     Use entity or multiple entities
-    //     /// </summary>
-    //     /// <param name="entityId">Unique id of the entity provided</param>
-    //     ITimerEntity Entity(params string[] entityId);
-    // }
-
-    // /// <summary>
-    // ///     The actions to perform on time functions
-    // /// </summary>
-    // public interface ITimerAction
-    // {
-    //     /// <summary>
-    //     ///     Excecute commands
-    //     /// </summary>
-    //     ISchedulerResult Execute();
-
-    //     /// <summary>
-    //     ///     Use attribute on entity action in time actions
-    //     /// </summary>
-    //     /// <param name="name">Name of attribute</param>
-    //     /// <param name="value">Value of attribute</param>
-    //     ITimerAction UsingAttribute(string name, object value);
-    // }
-
-    // /// <summary>
-    // ///     Entity timer actions
-    // /// </summary>
-    // public interface ITimerEntity : ITurnOff<ITimerAction>, ITurnOn<ITimerAction>, IToggle<ITimerAction>
-    // {
-    // }
 
     #region Media, Play, Stop, Pause, PlayPause, Speak
 
@@ -391,672 +327,4 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
     }
 
     #endregion Entities, TurnOn, TurnOff, Toggle
-
-    /// <summary>
-    ///     Base class for entity fluent types
-    /// </summary>
-    public class EntityBase : EntityState
-    {
-        internal readonly ConcurrentQueue<FluentAction> _actions =
-            new ConcurrentQueue<FluentAction>();
-
-        /// <summary>
-        ///     The daemon used in the API
-        /// </summary>
-        protected readonly INetDaemon Daemon;
-
-        /// <summary>
-        ///     The daemon used in the API
-        /// </summary>
-        protected readonly INetDaemonApp App;
-
-        /// <summary>
-        ///     The EntityIds used
-        /// </summary>
-        protected readonly IEnumerable<string> EntityIds;
-
-        internal FluentAction? _currentAction;
-
-
-        internal StateChangedInfo _currentState = new StateChangedInfo();
-
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        /// <param name="entityIds">The unique ids of the entities managed</param>
-        /// <param name="daemon">The Daemon that will handle API calls to Home Assistant</param>
-        /// <param name="app">The Daemon App calling fluent API</param>
-        public EntityBase(IEnumerable<string> entityIds, INetDaemon daemon, INetDaemonApp app)
-        {
-            EntityIds = entityIds;
-            Daemon = daemon;
-            App = app;
-        }
-
-        /// <inheritdoc/>
-        protected async Task CallServiceOnAllEntities(string service, dynamic? serviceData = null)
-        {
-            var taskList = new List<Task>();
-            foreach (var entityId in EntityIds)
-            {
-                var domain = GetDomainFromEntity(entityId);
-                serviceData ??= new FluentExpandoObject();
-                serviceData.entity_id = entityId;
-                var task = Daemon.CallService(domain, service, serviceData);
-                taskList.Add(task);
-            }
-
-            if (taskList.Count > 0) await Task.WhenAny(Task.WhenAll(taskList.ToArray()), Task.Delay(5000)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc/>
-        protected static string GetDomainFromEntity(string entity)
-        {
-            var entityParts = entity.Split('.');
-            if (entityParts.Length != 2)
-                throw new ApplicationException($"entity_id is mal formatted {entity}");
-
-            return entityParts[0];
-        }
-    }
-
-    /// <summary>
-    ///     Implements interface for managing entities in the fluent API
-    /// </summary>
-    public class EntityManager : EntityBase, IEntity, IAction,
-        IStateEntity, IState, IStateAction, IScript, IDelayStateChange
-    {
-
-
-
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        /// <param name="entityIds">The unique ids of the entities managed</param>
-        /// <param name="daemon">The Daemon that will handle API calls to Home Assistant</param>
-        /// <param name="app">The Daemon App calling fluent API</param>
-        public EntityManager(IEnumerable<string> entityIds, INetDaemon daemon, INetDaemonApp app) : base(entityIds, daemon, app)
-        {
-
-        }
-
-        /// <inheritdoc/>
-        public IState AndNotChangeFor(TimeSpan timeSpan)
-        {
-            _currentState.ForTimeSpan = timeSpan;
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IExecute Call(Func<string, EntityState?, EntityState?, Task> func)
-        {
-            _currentState.FuncToCall = func;
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public void Execute()
-        {
-            foreach (var entityId in EntityIds)
-                Daemon.ListenState(entityId, async (entityIdInn, newState, oldState) =>
-                {
-                    try
-                    {
-                        var entityManager = (EntityManager)_currentState.Entity!;
-
-                        if (_currentState.Lambda != null)
-                        {
-                            try
-                            {
-                                if (!_currentState.Lambda(newState, oldState))
-                                    return;
-                            }
-                            catch (Exception e)
-                            {
-                                Daemon.Logger.LogWarning(e,
-                                    "Failed to evaluate function in App {appId}, EntityId: {entityId}, From: {newState} To: {oldState}", App.Id, entityIdInn, $"{newState?.State}", $"{oldState?.State}");
-                                return;
-                            }
-                        }
-                        else
-                        {
-                            if (_currentState.To != null)
-                                if (_currentState.To != newState?.State)
-                                    return;
-
-                            if (_currentState.From != null)
-                                if (_currentState.From != oldState?.State)
-                                    return;
-
-                            // If we don´t accept all changes in the state change
-                            // and we do not have a state change so return
-                            if (newState?.State == oldState?.State && !_currentState.AllChanges)
-                                return;
-                        }
-
-                        if (_currentState.ForTimeSpan != TimeSpan.Zero)
-                        {
-                            Daemon.Logger.LogDebug(
-                                "AndNotChangeFor statement found, delaying {time}", _currentState.ForTimeSpan);
-                            await Task.Delay(_currentState.ForTimeSpan).ConfigureAwait(false);
-                            var currentState = Daemon.GetState(entityIdInn);
-                            if (currentState != null && currentState.State == newState?.State)
-                            {
-                                //var timePassed = newState.LastChanged.Subtract(currentState.LastChanged);
-                                if (currentState?.LastChanged == newState?.LastChanged)
-                                {
-                                    // No state has changed during the period
-                                    Daemon.Logger.LogDebug(
-                                        "State same {newState} during period of {time}, executing action!", $"{newState?.State}", _currentState.ForTimeSpan);
-                                    // The state has not changed during the time we waited
-                                    if (_currentState.FuncToCall == null)
-                                        await entityManager.ExecuteAsync(true).ConfigureAwait(false);
-                                    else
-                                    {
-                                        try
-                                        {
-                                            await _currentState.FuncToCall(entityIdInn, newState, oldState).ConfigureAwait(false);
-                                        }
-                                        catch (Exception e)
-                                        {
-                                            Daemon.Logger.LogWarning(e,
-                                                "Call function error in timespan in App {appId}, EntityId: {entityId}, From: {newState} To: {oldState}",
-                                                    App.Id, entityIdInn, $"{newState?.State}", $"{oldState?.State}");
-                                        }
-                                    }
-
-                                }
-                                else
-                                {
-                                    Daemon.Logger.LogDebug(
-                                        "State same {newState} but different state changed: {currentLastChanged}, expected {newLastChanged}",
-                                            $"{newState?.State}",
-                                            currentState?.LastChanged,
-                                            newState?.LastChanged);
-                                }
-                            }
-                            else
-                            {
-                                Daemon.Logger.LogDebug(
-                                    "State not same, do not execute for statement. {newState} found, expected {currentState}",
-                                    $"{newState?.State}",
-                                    $"{currentState?.State}");
-                            }
-                        }
-                        else
-                        {
-                            Daemon.Logger.LogDebug(
-                                "State {newState} expected from {oldState}, executing action!",
-                                    $"{newState?.State}",
-                                    $"{oldState?.State}"
-                                    );
-
-                            if (_currentState.FuncToCall != null)
-                            {
-                                try
-                                {
-                                    await _currentState.FuncToCall(entityIdInn, newState, oldState).ConfigureAwait(false);
-                                }
-                                catch (Exception e)
-                                {
-                                    Daemon.Logger.LogWarning(e,
-                                               "Call function error in App {appId}, EntityId: {entityId}, From: {newState} To: {oldState}",
-                                                   App.Id, entityIdInn, $"{newState?.State}", $"{oldState?.State}");
-
-                                }
-                            }
-
-                            else if (_currentState.ScriptToCall != null)
-                                await Daemon.RunScript(App, _currentState.ScriptToCall).ExecuteAsync().ConfigureAwait(false);
-                            else
-                                await entityManager.ExecuteAsync(true).ConfigureAwait(false);
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        Daemon.Logger.LogWarning(e, "Unhandled error in ListenState in App {appId}", App.Id);
-                    }
-
-                });
-
-
-            //}
-        }
-
-        /// <inheritdoc/>
-        IAction ISetState<IAction>.SetState(dynamic state)
-        {
-            _currentAction = new FluentAction(FluentActionType.SetState);
-            _currentAction.State = state;
-            _actions.Enqueue(_currentAction);
-            return this;
-        }
-
-        /// <inheritdoc/>
-        IStateAction ISetState<IStateAction>.SetState(dynamic state)
-        {
-            _currentState?.Entity?.SetState(state);
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IAction Toggle()
-        {
-            _currentAction = new FluentAction(FluentActionType.Toggle);
-            _actions.Enqueue(_currentAction);
-            return this;
-        }
-
-        /// <inheritdoc/>
-        IStateAction IToggle<IStateAction>.Toggle()
-        {
-            _currentState?.Entity?.Toggle();
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IAction TurnOff()
-        {
-            _currentAction = new FluentAction(FluentActionType.TurnOff);
-            _actions.Enqueue(_currentAction);
-            return this;
-        }
-
-        /// <inheritdoc/>
-        IStateAction ITurnOff<IStateAction>.TurnOff()
-        {
-            _currentState?.Entity?.TurnOff();
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IAction TurnOn()
-        {
-            _currentAction = new FluentAction(FluentActionType.TurnOn);
-            _actions.Enqueue(_currentAction);
-            return this;
-        }
-
-        /// <inheritdoc/>
-        IDelayResult IDelayStateChange.DelayUntilStateChange(object? to, object? from, bool allChanges)
-        {
-            return this.Daemon.DelayUntilStateChange(this.EntityIds, to, from, allChanges);
-        }
-
-        /// <inheritdoc/>
-        IDelayResult IDelayStateChange.DelayUntilStateChange(Func<EntityState?, EntityState?, bool> stateFunc)
-        {
-            return this.Daemon.DelayUntilStateChange(this.EntityIds, stateFunc);
-        }
-
-        IStateAction ITurnOn<IStateAction>.TurnOn()
-        {
-            _currentState?.Entity?.TurnOn();
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IStateEntity UseEntities(Func<IEntityProperties, bool> func)
-        {
-            _currentState.Entity = Daemon.Entities(App, func);
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IStateEntity UseEntities(IEnumerable<string> entities)
-        {
-            _currentState.Entity = Daemon.Entities(App, entities);
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IStateEntity UseEntity(params string[] entityId)
-        {
-            _currentState.Entity = Daemon.Entity(App, entityId);
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IState WhenStateChange(object? to = null, object? from = null, bool allChanges = false)
-        {
-            _currentState = new StateChangedInfo
-            {
-                From = from,
-                To = to,
-                AllChanges = allChanges
-            };
-
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IState WhenStateChange(Func<EntityState?, EntityState?, bool> stateFunc)
-        {
-            _currentState = new StateChangedInfo
-            {
-                Lambda = stateFunc
-            };
-            return this;
-        }
-
-        /// <inheritdoc/>
-        public IAction WithAttribute(string name, object value)
-        {
-            if (_currentAction != null) _currentAction.Attributes[name] = value;
-            return this;
-        }
-
-        /// <inheritdoc/>
-        IStateAction IStateAction.WithAttribute(string name, object value)
-        {
-            var entityManager = (EntityManager)_currentState.Entity!;
-            entityManager.WithAttribute(name, value);
-
-            return this;
-        }
-
-        /// <inheritdoc/>
-        Task IExecuteAsync.ExecuteAsync()
-        {
-            return ExecuteAsync();
-        }
-
-        /// <inheritdoc/>
-        async Task IScript.ExecuteAsync()
-        {
-            var taskList = new List<Task>();
-            foreach (var scriptName in EntityIds)
-            {
-                var name = scriptName;
-                if (scriptName.Contains('.'))
-                    name = scriptName[(scriptName.IndexOf('.') + 1)..];
-                var task = Daemon.CallService("script", name);
-                taskList.Add(task);
-            }
-
-            // Wait for all tasks to complete or max 5 seconds
-            if (taskList.Count > 0) await Task.WhenAny(Task.WhenAll(taskList.ToArray()), Task.Delay(5000)).ConfigureAwait(false);
-        }
-
-        /// <summary>
-        ///     Executes the sequence of actions
-        /// </summary>
-        /// <param name="keepItems">
-        ///     True if  you want to keep items
-        /// </param>
-        /// <remarks>
-        ///     You want to keep the items when using this as part of an automation
-        ///     that are kept over time. Not keeping when just doing a command
-        /// </remarks>
-        /// <returns></returns>
-        public async Task ExecuteAsync(bool keepItems)
-        {
-            if (keepItems)
-                foreach (var action in _actions)
-                    await HandleAction(action).ConfigureAwait(false);
-            else
-                while (_actions.TryDequeue(out var fluentAction))
-                    await HandleAction(fluentAction).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc/>
-        public IExecute RunScript(params string[] entityIds)
-        {
-            _currentState.ScriptToCall = entityIds;
-            return this;
-        }
-
-        /// <inheritdoc/>
-        private Task ExecuteAsync()
-        {
-            return ExecuteAsync(false);
-        }
-
-        /// <inheritdoc/>
-        private async Task HandleAction(FluentAction fluentAction)
-        {
-            var attributes = fluentAction.Attributes.Select(n => (n.Key, n.Value)).ToArray();
-
-            var taskList = new List<Task>();
-            foreach (var entityId in EntityIds)
-            {
-                var task = fluentAction.ActionType switch
-                {
-                    FluentActionType.TurnOff => TurnOffAsync(entityId, attributes),
-                    FluentActionType.TurnOn => TurnOnAsync(entityId, attributes),
-                    FluentActionType.Toggle => ToggleAsync(entityId, attributes),
-                    FluentActionType.SetState => Daemon.SetState(entityId, fluentAction.State, attributes),
-                    _ => throw new NotSupportedException($"Fluent action type not handled! {fluentAction.ActionType}")
-                };
-                taskList.Add(task);
-            }
-            // Wait for all tasks to complete or max 5 seconds
-            if (taskList.Count > 0) await Task.WhenAny(Task.WhenAll(taskList.ToArray()), Task.Delay(5000)).ConfigureAwait(false);
-        }
-
-        private Task ToggleAsync(string entityId, params (string name, object val)[] attributeNameValuePair)
-        {
-            // Get the domain if supported, else domain is homeassistant
-            string domain = GetDomainFromEntity(entityId);
-            // Use it if it is supported else use default "homeassistant" domain
-
-            // Use expando object as all other methods
-            dynamic attributes = attributeNameValuePair.ToDynamic();
-            // and add the entity id dynamically
-            attributes.entity_id = entityId;
-
-            return Daemon.CallService(domain, "toggle", attributes, false);
-        }
-
-        private Task TurnOffAsync(string entityId, params (string name, object val)[] attributeNameValuePair)
-        {
-            // Get the domain if supported, else domain is homeassistant
-            string domain = GetDomainFromEntity(entityId);
-            // Use it if it is supported else use default "homeassistant" domain
-
-            // Use expando object as all other methods
-            dynamic attributes = attributeNameValuePair.ToDynamic();
-            // and add the entity id dynamically
-            attributes.entity_id = entityId;
-
-            return Daemon.CallService(domain, "turn_off", attributes, false);
-        }
-
-        private Task TurnOnAsync(string entityId, params (string name, object val)[] attributeNameValuePair)
-        {
-            // Use default domain "homeassistant" if supported is missing
-            string domain = GetDomainFromEntity(entityId);
-            // Use it if it is supported else use default "homeassistant" domain
-
-            // Convert the value pairs to dynamic type
-            dynamic attributes = attributeNameValuePair.ToDynamic();
-            // and add the entity id dynamically
-            attributes.entity_id = entityId;
-
-            return Daemon.CallService(domain, "turn_on", attributes, false);
-        }
-    }
-
-    // /// <summary>
-    // ///     Handles timer features
-    // /// </summary>
-    // public class TimeManager : ITime, ITimeItems, ITimerEntity, ITimerAction
-    // {
-    //     private readonly INetDaemon _daemon;
-    //     private readonly INetDaemonApp _app;
-    //     private readonly List<string> _entityIds = new List<string>();
-    //     private TimeSpan _timeSpan;
-    //     private EntityManager? _entityManager;
-
-    //     /// <summary>
-    //     ///     Constructor
-    //     /// </summary>
-    //     /// <param name="daemon">The NetDaemon that handles API calls to Home Assistant</param>
-    //     /// <param name="app">The Daemon App calling fluent API</param>
-    //     public TimeManager(INetDaemon daemon, INetDaemonApp app)
-    //     {
-    //         _daemon = daemon;
-    //         _app = app;
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ITimerEntity Entities(Func<IEntityProperties, bool> func)
-    //     {
-    //         var x = _daemon.State.Where(func);
-    //         _entityIds.AddRange(x.Select(n => n.EntityId).ToArray());
-    //         _entityManager = new EntityManager(_entityIds.ToArray(), _daemon, _app);
-    //         return this;
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ITimerEntity Entities(IEnumerable<string> entities)
-    //     {
-    //         _entityIds.AddRange(entities);
-    //         _entityManager = new EntityManager(_entityIds, _daemon, _app);
-    //         return this;
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ITimerEntity Entity(params string[] entityIds)
-    //     {
-    //         _entityIds.AddRange(entityIds);
-    //         _entityManager = new EntityManager(entityIds, _daemon, _app);
-    //         return this;
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ITimeItems Every(TimeSpan timeSpan)
-    //     {
-    //         _timeSpan = timeSpan;
-    //         return this;
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ISchedulerResult Execute()
-    //     {
-    //         _ = _entityManager ?? throw new NullReferenceException($"{nameof(_entityManager)} can not be null here!");
-
-    //         return _daemon.Scheduler.RunEvery(_timeSpan, async () => { await _entityManager.ExecuteAsync(true).ConfigureAwait(false); });
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ITimerAction Toggle()
-    //     {
-    //         _ = _entityManager ?? throw new NullReferenceException($"{nameof(_entityManager)} can not be null here!");
-
-    //         _entityManager.Toggle();
-    //         return this;
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ITimerAction TurnOff()
-    //     {
-    //         _ = _entityManager ?? throw new NullReferenceException($"{nameof(_entityManager)} can not be null here!");
-
-    //         _entityManager.TurnOff();
-    //         return this;
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ITimerAction TurnOn()
-    //     {
-    //         _ = _entityManager ?? throw new NullReferenceException($"{nameof(_entityManager)} can not be null here!");
-
-    //         _entityManager.TurnOn();
-    //         return this;
-    //     }
-
-    //     /// <inheritdoc/>
-    //     public ITimerAction UsingAttribute(string name, object value)
-    //     {
-    //         _ = _entityManager ?? throw new NullReferenceException($"{nameof(_entityManager)} can not be null here!");
-
-    //         _entityManager.WithAttribute(name, value);
-    //         return this;
-    //     }
-    // }
-
-    /// <summary>
-    ///     Represents data about an action in a fluent API
-    /// </summary>
-    internal class FluentAction
-    {
-        // Todo: refactor the action class to manage only data that is relevant
-
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        /// <param name="type">The type of action that is being managed</param>
-        public FluentAction(FluentActionType type)
-        {
-            ActionType = type;
-            Attributes = new Dictionary<string, object>();
-        }
-
-        /// <summary>
-        ///     Type of action
-        /// </summary>
-        public FluentActionType ActionType { get; }
-
-        /// <summary>
-        ///     Attributes used in action if specified
-        /// </summary>
-        public Dictionary<string, object> Attributes { get; }
-
-        /// <summary>
-        ///     Message to speak if it is a speak action
-        /// </summary>
-        public string? MessageToSpeak { get; internal set; }
-
-        /// <summary>
-        ///     The state to manage for state actions
-        /// </summary>
-        public dynamic? State { get; set; }
-    }
-
-    /// <summary>
-    ///     Information about state changed actions
-    /// </summary>
-    internal class StateChangedInfo
-    {
-        /// <summary>
-        ///     All changes tracked, even if only attributes
-        /// </summary>
-        public bool AllChanges { get; set; }
-
-        /// <summary>
-        ///     Entity changed
-        /// </summary>
-        public IEntity? Entity { get; set; }
-
-        /// <summary>
-        ///     Timespan it have kept same state
-        /// </summary>
-        public TimeSpan ForTimeSpan { get; set; }
-
-        /// <summary>
-        ///     From state
-        /// </summary>
-        public dynamic? From { get; set; }
-
-        /// <summary>
-        ///     To state
-        /// </summary>
-        public dynamic? To { get; set; }
-
-        /// <summary>
-        ///     Function to call when state changes
-        /// </summary>
-        public Func<string, EntityState?, EntityState?, Task>? FuncToCall { get; set; }
-
-        /// <summary>
-        ///     Filter state changes with lamda
-        /// </summary>
-        public Func<EntityState?, EntityState?, bool>? Lambda { get; set; }
-
-        /// <summary>
-        ///     Script to call when state changes
-        /// </summary>
-        public string[]? ScriptToCall { get; internal set; }
-    }
 }
