+++ src/DaemonRunner/DaemonRunner/HassioConfig.cs
@@ -2,7 +2,6 @@ using System.Text.Json.Serialization;
 
 public class HassioConfig
 {
-
     [JsonPropertyName("log_level")]
     public string? LogLevel { get; set; }
