+++ tests/NetDaemon.Daemon.Tests/DaemonRunner/Fixtures/dependtests/global_app.cs
@@ -33,8 +33,6 @@ namespace DependNs
 
             return Task.CompletedTask;
         }
-
-
     }
 
     public class GlobalApp : NetDaemonApp
