+++ tests/NetDaemon.Daemon.Tests/NetDaemonApp/AppExtensionsTests.cs
@@ -1,12 +1,10 @@
-using Xunit;
 using JoySoftware.HomeAssistant.NetDaemon.Common;
-using System;
+using Xunit;
 
 namespace NetDaemon.Daemon.Tests.NetDaemonApp
 {
     public class AppExtensionsTests
     {
-
         [Theory]
         [InlineData("ABCDEFGHIJKLMNOPQRSTUVXYZ", "abcdefghijklmnopqrstuvxyz")]
         [InlineData("Å_Ä_Ö", "a_a_o")]
