+++ tests/NetDaemon.Daemon.Tests/NetDaemonApp/FaultyAppsTests.cs
@@ -1,14 +1,5 @@
-using JoySoftware.HomeAssistant.NetDaemon.Common;
-using JoySoftware.HomeAssistant.NetDaemon.Daemon;
-using JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.App;
-using JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.Config;
 using Microsoft.Extensions.Logging;
 using Moq;
-using System;
-using System.Collections.Generic;
-using System.Dynamic;
-using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 
@@ -21,12 +12,11 @@ namespace NetDaemon.Daemon.Tests.NetDaemonApp
         private readonly JoySoftware.HomeAssistant.NetDaemon.Common.NetDaemonApp _app;
 
         public JoySoftware.HomeAssistant.NetDaemon.Common.NetDaemonApp App => _app;
+
         public FaultyAppTests() : base()
         {
-
             _app = new DaemonAppTestApp();
             _app.StartUpAsync(DefaultDaemonHost);
-
         }
 
         [Fact]
@@ -50,6 +40,5 @@ namespace NetDaemon.Daemon.Tests.NetDaemonApp
 
             LoggerMock.AssertLogged(LogLevel.Warning, Times.Once());
         }
-
     }
 }
